fix(amp): 修复 CUDA+AMP 路径下 VQ 与 eval 的 fp32/Half dtype 不匹配
三处仅在 CUDA 混合精度下触发(CPU/mock 单测不覆盖),导致训练/评估直接崩:
- VectorQuantizer EMA 加权和:encodings(fp32) 与 autocast 下 Half 的 z_flattened
  matmul dtype 冲突 → 改用已算好的 fp32 版 z_flattened_f32(buffer 本为 fp32,符合 I4);
  random restart 同步改用 fp32 版以匹配 embed/ema_w buffer。
- VQ 直连估计器(STE):量化结果在 fp32 buffer 域算出为 fp32,z + (z_q - z).detach()
  把 Half 输入的梯度提升为 fp32,反向报 "Found dtype Float but expected Half"
  → 量化结果 .to(z.dtype) 回到输入 dtype。
- eval.evaluate:预测器前向在 autocast 块外(fp32 权重),而 z 经 adapter Linear 在
  autocast 内为 Half → z/z_tgt 统一转 fp32(no_grad 诊断路径,与下游相关性统计同口径)。

验证:tests/unit+integration 24 passed;runs/data/micro 上 CUDA 训练+eval 端到端跑通。

+++ train/minecraft/eval.py
@@ -76,8 +76,10 @@ def evaluate(model, effect_tok, loader, device, amp_dev, use_amp, cfg):
             feats = model.extract_feats(img.reshape(B * T, *img.shape[2:]))
             z, _ = model.encode(feats)
         M = z.shape[-2]
-        z = z.view(B, T, M, model.d)
-        z_tgt = model.encode_target(feats).view(B, T, M, model.d)
+        # autocast 下 adapter 的 Linear 产出 Half;eval 的预测器前向在 autocast 块外
+        # (fp32 权重),且下游 z_tgt/相关性统计均为 fp32 —— 统一转 fp32 避免 dtype 不匹配。
+        z = z.view(B, T, M, model.d).float()
+        z_tgt = model.encode_target(feats).view(B, T, M, model.d).float()
 
         tf = torch.cat([torch.zeros(B, 1, device=dt.device), dt.cumsum(dim=1)], dim=1)
         target = T - 1
